+++ src/pymovements/events/__init__.py
@@ -67,7 +67,7 @@ from pymovements.events.detection import ivt
 from pymovements.events.detection import microsaccades
 from pymovements.events.detection._library import EventDetectionLibrary
 from pymovements.events.detection._library import register_event_detection
-from pymovements.events.frame import Events
+from pymovements.events.events import Events
 from pymovements.events.precomputed import PrecomputedEventDataFrame
 from pymovements.events.processing import EventGazeProcessor
 from pymovements.events.processing import EventProcessor
