+++ src/pymovements/events/detection/_idt.py
@@ -26,7 +26,7 @@ from pymovements._utils import _checks
 from pymovements.events._utils._filters import events_split_nans
 from pymovements.events._utils._filters import filter_candidates_remove_nans
 from pymovements.events.detection._library import register_event_detection
-from pymovements.events.frame import Events
+from pymovements.events.events import Events
 
 
 def dispersion(positions: list[list[float]] | np.ndarray) -> float:
