+++ src/pymovements/events/detection/_ivt.py
@@ -25,7 +25,7 @@ import numpy as np
 from pymovements._utils import _checks
 from pymovements.events._utils._filters import filter_candidates_remove_nans
 from pymovements.events.detection._library import register_event_detection
-from pymovements.events.frame import Events
+from pymovements.events.events import Events
 from pymovements.gaze.transforms_numpy import consecutive
 from pymovements.gaze.transforms_numpy import norm
